feat(clerk-react): Resolve PR comments
