+++ packages/clerk-js/src/core/resources/EmailAddress.ts
@@ -32,14 +32,18 @@ export class EmailAddress extends BaseResource implements EmailAddressResource {
     });
   }
 
-  prepareVerification = (params?: PrepareEmailAddressVerificationParams): Promise<this> => {
+  prepareVerification = (
+    params?: PrepareEmailAddressVerificationParams,
+  ): Promise<this> => {
     return this._basePost<EmailAddressJSON>({
       action: 'prepare_verification',
       body: { ...(params || { strategy: 'email_code' }) },
     });
   };
 
-  attemptVerification = (params: AttemptEmailAddressVerificationParams): Promise<this> => {
+  attemptVerification = (
+    params: AttemptEmailAddressVerificationParams,
+  ): Promise<this> => {
     const { code } = params || {};
     return this._basePost<EmailAddressJSON>({
       action: 'attempt_verification',
@@ -47,10 +51,15 @@ export class EmailAddress extends BaseResource implements EmailAddressResource {
     });
   };
 
-  createMagicLinkFlow = (): CreateMagicLinkFlowReturn<StartMagicLinkFlowParams, EmailAddressResource> => {
+  createMagicLinkFlow = (): CreateMagicLinkFlowReturn<
+    StartMagicLinkFlowParams,
+    EmailAddressResource
+  > => {
     const { run, stop } = Poller();
 
-    const startMagicLinkFlow = async ({ redirectUrl }: StartMagicLinkFlowParams): Promise<EmailAddressResource> => {
+    const startMagicLinkFlow = async ({
+      redirectUrl,
+    }: StartMagicLinkFlowParams): Promise<EmailAddressResource> => {
       if (!this.id) {
         clerkVerifyEmailAddressCalledBeforeCreate('SignUp');
       }
@@ -89,7 +98,9 @@ export class EmailAddress extends BaseResource implements EmailAddressResource {
     this.id = data.id;
     this.emailAddress = data.email_address;
     this.verification = new Verification(data.verification);
-    this.linkedTo = (data.linked_to || []).map(link => new IdentificationLink(link));
+    this.linkedTo = (data.linked_to || []).map(
+      link => new IdentificationLink(link),
+    );
     return this;
   }
 }
