+++ packages/clerk-js/src/core/resources/PhoneNumber.ts
@@ -1,10 +1,11 @@
 import type {
+  AttemptPhoneNumberVerificationParams,
   IdentificationLinkResource,
   PhoneNumberJSON,
   PhoneNumberResource,
+  SetReservedForSecondFactorParams,
   VerificationResource,
 } from '@clerk/types';
-import { AttemptPhoneNumberVerificationParams, SetReservedForSecondFactorParams } from '@clerk/types/src';
 
 import { BaseResource, IdentificationLink, Verification } from './internal';
 
@@ -35,14 +36,18 @@ export class PhoneNumber extends BaseResource implements PhoneNumberResource {
     });
   };
 
-  attemptVerification = (params: AttemptPhoneNumberVerificationParams): Promise<PhoneNumberResource> => {
+  attemptVerification = (
+    params: AttemptPhoneNumberVerificationParams,
+  ): Promise<PhoneNumberResource> => {
     const { code } = params || {};
     return this._basePost<PhoneNumberJSON>({
       action: 'attempt_verification',
       body: { code },
     });
   };
-  setReservedForSecondFactor = (params: SetReservedForSecondFactorParams): Promise<PhoneNumberResource> => {
+  setReservedForSecondFactor = (
+    params: SetReservedForSecondFactorParams,
+  ): Promise<PhoneNumberResource> => {
     const { reserved } = params || {};
     return this._basePatch<PhoneNumberJSON>({
       body: { reserved_for_second_factor: reserved },
@@ -74,7 +79,9 @@ export class PhoneNumber extends BaseResource implements PhoneNumberResource {
     this.reservedForSecondFactor = data.reserved_for_second_factor;
     this.defaultSecondFactor = data.default_second_factor;
     this.verification = new Verification(data.verification);
-    this.linkedTo = (data.linked_to || []).map(link => new IdentificationLink(link));
+    this.linkedTo = (data.linked_to || []).map(
+      link => new IdentificationLink(link),
+    );
     return this;
   }
 }
