+++ packages/types/src/externalAccount.ts
@@ -1,4 +1,4 @@
-import { OAuthProvider } from './providers';
+import { OAuthProvider } from './oauth';
 
 export interface ExternalAccountResource {
   id: string;
