+++ packages/types/src/oauth.ts
@@ -1,4 +1,3 @@
-import { OAuthProvider } from './providers';
 import { OAuthStrategy } from './strategies';
 
 export interface OAuthProviderData {
@@ -7,6 +6,39 @@ export interface OAuthProviderData {
   name: string;
   docsUrl: string;
 }
+
+export type FacebookOauthProvider = 'facebook';
+export type GoogleOauthProvider = 'google';
+export type HubspotOauthProvider = 'hubspot';
+export type GithubOauthProvider = 'github';
+export type TiktokOauthProvider = 'tiktok';
+export type GitlabOauthProvider = 'gitlab';
+export type DiscordOauthProvider = 'discord';
+export type TwitterOauthProvider = 'twitter';
+export type TwitchOauthProvider = 'twitch';
+export type LinkedinOauthProvider = 'linkedin';
+export type DropboxOauthProvider = 'dropbox';
+export type BitbucketOauthProvider = 'bitbucket';
+export type MicrosoftOauthProvider = 'microsoft';
+export type NotionOauthProvider = 'notion';
+export type AppleOauthProvider = 'apple';
+
+export type OAuthProvider =
+  | FacebookOauthProvider
+  | GoogleOauthProvider
+  | HubspotOauthProvider
+  | GithubOauthProvider
+  | TiktokOauthProvider
+  | GitlabOauthProvider
+  | DiscordOauthProvider
+  | TwitterOauthProvider
+  | TwitchOauthProvider
+  | LinkedinOauthProvider
+  | DropboxOauthProvider
+  | BitbucketOauthProvider
+  | MicrosoftOauthProvider
+  | NotionOauthProvider
+  | AppleOauthProvider;
 
 export const OAUTH_PROVIDERS: OAuthProviderData[] = [
   {
