--- packages/types/src/providers.ts
@@ -1,36 +1,0 @@
-export type FacebookOauthProvider = 'facebook';
-export type GoogleOauthProvider = 'google';
-export type HubspotOauthProvider = 'hubspot';
-export type GithubOauthProvider = 'github';
-export type TiktokOauthProvider = 'tiktok';
-export type GitlabOauthProvider = 'gitlab';
-export type DiscordOauthProvider = 'discord';
-export type TwitterOauthProvider = 'twitter';
-export type TwitchOauthProvider = 'twitch';
-export type LinkedinOauthProvider = 'linkedin';
-export type DropboxOauthProvider = 'dropbox';
-export type BitbucketOauthProvider = 'bitbucket';
-export type MicrosoftOauthProvider = 'microsoft';
-export type NotionOauthProvider = 'notion';
-export type AppleOauthProvider = 'apple';
-
-export type OAuthProvider =
-  | FacebookOauthProvider
-  | GoogleOauthProvider
-  | HubspotOauthProvider
-  | GithubOauthProvider
-  | TiktokOauthProvider
-  | GitlabOauthProvider
-  | DiscordOauthProvider
-  | TwitterOauthProvider
-  | TwitchOauthProvider
-  | LinkedinOauthProvider
-  | DropboxOauthProvider
-  | BitbucketOauthProvider
-  | MicrosoftOauthProvider
-  | NotionOauthProvider
-  | AppleOauthProvider;
-
-export type MetamaskWeb3Provider = 'metamask';
-
-export type Web3Provider = MetamaskWeb3Provider;
