+++ packages/types/src/strategies.ts
@@ -1,4 +1,5 @@
-import { OAuthProvider, Web3Provider } from './providers';
+import { OAuthProvider } from './oauth';
+import { Web3Provider } from './web3';
 
 export type PasswordStrategy = 'password';
 export type PhoneCodeStrategy = 'phone_code';
