+++ packages/types/src/web3.ts
@@ -1,0 +1,3 @@
+export type MetamaskWeb3Provider = 'metamask';
+
+export type Web3Provider = MetamaskWeb3Provider;
